TOR-23 added bet badge to show the current bet, moved colors to constants
