+++ src/components/Button/style.module.scss
@@ -17,9 +17,9 @@
 }
 
 .buttonRed {
-  background-color: #ff0031;
+  background-color: rgb(var(--red));
 }
 
 .buttonGreen {
-  background-color: #1dbe45;
+  background-color: rgb(var(--green));
 }
