+++ src/components/Roulette/Bet/Badge/style.module.scss
@@ -1,0 +1,12 @@
+.badge {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-width: 40px;
+  width: 100%;
+  height: 40px;
+  padding: 10px;
+  margin-right: 15px;
+  border-radius: 14px;
+  user-select: none;
+}
