+++ src/components/Roulette/Bet/style.module.scss
@@ -1,6 +1,10 @@
 .betContainer {
   display: flex;
+  align-items: center;
 }
+
 .betSelector {
   outline: none;
+  height: 40px;
+  border-radius: 5px;
 }
