+++ src/components/Roulette/LastResults/style.module.scss
@@ -12,13 +12,3 @@
   margin-left: 10px;
   border-radius: 50%;
 }
-
-.red {
-  background-color: #ff0031;
-}
-.gray {
-  background-color: #474747;
-}
-.green {
-  background-color: #1dbe45;
-}
